backend/requirements.txt
Django==2.2.4
djangorestframework==3.10.2
joblib==0.13.2
numpy==1.17.0
pytz==2019.2
scikit-learn==0.21.3
scipy==1.3.0
sklearn==0.0
sqlparse==0.3.0
django-cors-headers